+++ src/co/edu/unbosque/model/persistence/ApuestaDAO.java
@@ -40,7 +40,7 @@ public class ApuestaDAO implements CRUDOperation {
 	public void create(String... args) {
 		ApuestaDTO bet = new ApuestaDTO();
 		bet.setBookmakerOffice(args[0]);
-		bet.setIdentification(Double.parseDouble(args[1]));
+		bet.setIdentification(Long.parseLong(args[1]));
 		bet.setDayOfTheWeek(args[2]);
 		bet.setValueOfTheBet(Double.parseDouble(args[3]));
 		listOfBets.add(bet);
@@ -72,7 +72,7 @@ public class ApuestaDAO implements CRUDOperation {
 				listOfBets.get(index).setBookmakerOffice(args[0]);
 			}
 			if (!args[1].isBlank() || !args[1].isEmpty() || args[1] != null) {
-				listOfBets.get(index).setIdentification(Double.parseDouble(args[1]));
+				listOfBets.get(index).setIdentification(Long.parseLong(args[1]));
 			}
 			if (!args[2].isBlank() || !args[2].isEmpty() || args[2] != null) {
 				listOfBets.get(index).setDayOfTheWeek(args[2]);
